+++ src/components/memory/Memory.java
@@ -4,7 +4,7 @@ package components.memory;
  * Interface to a memory
  */
 public interface Memory {
-    public byte read(int address);
-    public void write(int address, byte value);
+    public int read(int address);
+    public void write(int address, int value);
     public MemoryRange getRange();
 }
